# claim-for-payment-service/generated/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: https://laa-data-stewardship-payments-dev.apps.live.cloud-platform.service.justice.gov.uk
tags:
- description: the claims API
  name: claims
paths:
  /api/v1/claims/{id}:
    delete:
      x-accepts:
      - application/json
      operationId: deleteClaim
      parameters:
      - description: The id of the claim to be deleted
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: Delete claim by id
      tags:
      - claims
    get:
      x-accepts:
      - application/json
      operationId: getClaimById
      parameters:
      - description: The id of the claim to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
          description: Internal server error
      summary: Get claim by id
      tags:
      - claims
    put:
      x-content-type: application/json
      x-accepts:
      - application/json
      operationId: updateClaim
      parameters:
      - description: The id of the claim to be updated
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequestBody"
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: Update an claim
      tags:
      - claims
  /api/v1/claims:
    get:
      x-accepts:
      - application/json
      operationId: getClaims
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Success
        "400":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Not found
        "500":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Claim"
                type: array
          description: Internal server error
      summary: Get claims
      tags:
      - claims
    post:
      x-content-type: application/json
      x-accepts:
      - application/json
      operationId: createClaim
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequestBody"
        required: true
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: Create an claim
      tags:
      - claims
components:
  schemas:
    ClaimRequestBody:
      description: create an claim
      example:
        name: name
        description: description
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - description
      - name
    Claim:
      example:
        name: name
        description: description
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        description:
          type: string
      required:
      - description
      - id
      - name